chore: Update folder structure
