+++ src/components/Card/styles.ts
@@ -1,9 +1,7 @@
 import styled, { css } from 'styled-components';
-import palette from '../theme/palette';
+import palette from '../../theme/palette';
 
-import { CardProps } from '../@types/Card';
-
-const Container = styled.article`
+export const Container = styled.article`
   background-color: ${palette.backgroundBase};
   border-radius: ${palette.borderRadius};
   box-shadow: ${palette.boxShadow1};
@@ -14,15 +12,13 @@ const Container = styled.article`
     transform: translate3d(0, -5px, 0);
   }
 `;
-
-const Figure = styled.figure`
+export const Figure = styled.figure`
   position: relative;
   height: 0;
   padding-bottom: 20%;
   background-color: #000;
 `;
-
-const Image = styled.img`
+export const Image = styled.img`
   position: absolute;
   top: 0;
   left: 0;
@@ -31,11 +27,9 @@ const Image = styled.img`
   object-fit: cover;
   opacity: 0.5;
 `;
-
-const Header = styled.header`
+export const Header = styled.header`
   padding: 10px;
 `;
-
 const gradientOverflow = css`
   &::after {
     content: '';
@@ -50,8 +44,7 @@ const gradientOverflow = css`
     pointer-events: none;
   }
 `;
-
-const Title = styled.h2`
+export const Title = styled.h2`
   position: relative;
   max-height: 3.75rem;
   margin-bottom: 10px;
@@ -65,8 +58,7 @@ const Title = styled.h2`
     height: 1.875rem;
   }
 `;
-
-const Description = styled.p`
+export const Description = styled.p`
   position: relative;
   max-height: 4.5rem;
   line-height: 1.5;
@@ -78,20 +70,3 @@ const Description = styled.p`
     height: 1.5rem;
   }
 `;
-
-// FIXME: Add tags when tag component is implemented
-export default function Card({ title, description, link, image }: CardProps) {
-  return (
-    <Container>
-      <a href={link} target="_blank" rel="noopener noreferrer">
-        <Figure>
-          <Image src={image} />
-        </Figure>
-        <Header>
-          <Title>{title}</Title>
-          <Description>{description}</Description>
-        </Header>
-      </a>
-    </Container>
-  );
-}
