+++ src/components/Card/types.ts
@@ -1,4 +1,4 @@
-import { Bookmark } from './domain';
+import { Bookmark } from '../../@types/domain';
 
 // FIXME: Fix the type below when defined
 export interface CardProps extends Bookmark {}
