+++ src/components/CardContainer/types.ts
@@ -1,4 +1,4 @@
-import { CardProps } from './Card';
+import { CardProps } from '../Card/types';
 
 export interface CardContainerProps {
   items: CardProps[];
